refactor: move buttons to shared folder
