+++ includes/publisher/js/src/components/shared/repeaters/DeleteItemButton.jsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx, css } from "@emotion/react";
-import TrashIcon from "../../../../../components/icons/TrashIcon";
-import { colors } from "../../../../../shared-assets/js/emotion";
+import TrashIcon from "../../../../../../components/icons/TrashIcon";
+import { colors } from "../../../../../../shared-assets/js/emotion";
 import { __ } from "@wordpress/i18n";
 
 const DeleteItemButton = ({ deleteItem }) => {
